Source Codes for the Raspberry Pi and Arduino
The python script Ball_Pos_Color_Detection.py uses a color detection algorithm to track the position of the ball. The Servo_Control.txt code receives the position of the ball and determines which direction the plate should tilt and by what angle, then sends the signal to the servo motors.